Improve performance, documentation, and modularity
