+++ neqo-transport/src/connection/mod.rs
@@ -18,9 +18,12 @@ use std::{
 };
 
 use neqo_common::{
-    Buffer, Datagram, Decoder, Ecn, Encoder, Hex, HexSnipMiddle, HexWithLen, Role, Tos, datagram,
-    event::Provider as EventProvider, hrtime, qdebug, qerror, qinfo, qlog::Qlog, qtrace, qwarn,
-    to_u64, to_usize,
+    Buffer, Datagram, Decoder, Ecn, Encoder, Role, Tos, datagram,
+    event::Provider as EventProvider,
+    hex::{Hex, HexSnipMiddle, HexWithLen},
+    hrtime, qdebug, qerror, qinfo,
+    qlog::Qlog,
+    qtrace, qwarn, to_u64, to_usize,
 };
 use nss::{
     Agent, AntiReplay, AuthenticationStatus, Cipher, Client, Group, HandshakeState, PrivateKey,
